move SpType and SpSeed into separate headers
